% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.ci.R
\name{alpha.ci}
\alias{alpha.ci}
\title{Confidence Interval Estimator}
\usage{
alpha.ci(per, sample)
}
\arguments{
\item{per}{Percentage of confidence}

\item{sample}{A vector of sample data}
}
\value{
A confidence interval for the given data
}
\description{
Confidence Interval Estimator
}
