% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mode.R
\name{find.mode}
\alias{find.mode}
\title{Finding the Mode of a Data set}
\usage{
find.mode(x)
}
\arguments{
\item{x}{}
}
\value{
the mode of a data set
}
\description{
Finding the Mode of a Data set
}
\examples{
find_mode(c(1,2,3,3,4,5))
}
