% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{find_mode}
\alias{find_mode}
\title{Finding the Mode of a Data set}
\usage{
find_mode(x)
}
\arguments{
\item{x}{}
}
\value{
the mode of a data set
}
\description{
Finding the Mode of a Data set
}
\examples{
find_mode(c(1,2,3,3,4,5))
}
