% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire}
\format{
a data frame with 15 rows 2 variables
 \describe{
   \item{DISTANCE}{}
   \item{DAMAGE}{}
   '''
 }
}
\usage{
fire
}
\description{
A data set being used for MATH 4753
}
\keyword{datasets}
