% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myNRML.R
\name{myNRML}
\alias{myNRML}
\title{myNRML}
\usage{
myNRML(x0, delta = 0.001, llik, xrange, parameter = "param")
}
\arguments{
\item{x0}{X-not first approximation value}

\item{delta}{increment, try 0.001}

\item{llik}{:}

\item{xrange}{:}

\item{parameter}{:}
}
\description{
This is a
}
\examples{
myNRML(x0=0.99,delta=0.000001,llik=function(x)
log(dbinom(12,size=20,prob=x)*dbinom(10,size=25,prob=x)),
xrange=c(0.01,0.99),parameter="p" )
}
