% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neg_bi_nom.R
\name{neg_bi_nom}
\alias{neg_bi_nom}
\title{Negative Binomial Probability}
\usage{
neg_bi_nom(y, r, p)
}
\arguments{
\item{y}{Number of trials until the rth success is observed}

\item{r}{rth success}

\item{p}{probability of success on a single Bernoulli trial}
}
\value{
The probability distribution for a Negative Binomial of random variable Y
}
\description{
Negative Binomial Probability
}
\examples{
#P(Y=10), Y~NegBin(p=0.4,r=3)
neg_bi_nom(10,3,0.4)
}
