% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.data.select.R
\name{rand.data.select}
\alias{rand.data.select}
\title{Random Data Selection}
\usage{
rand.data.select(data.object, n)
}
\arguments{
\item{data.object}{A data frame/csv}

\item{n}{Number of desired rows}
}
\value{
<<<<<<< HEAD
An object data.select, with the randomly selected data.
}
\description{
Selects a specific number of random rows from a list of data.
=======
An object data.select, with the randomly slected data.
}
\description{
Selects a specific numebr of random rows from a list of data.
>>>>>>> e87b82bcf5da1d3055578f791e96cf47d42ebf92
}
