% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Iterations}

\item{x}{Data vector}

\item{fun}{Function desired}

\item{alpha}{Confidence interval: (1-alpha)*100 = percent. 95\%-> alpha=0.05}

\item{cx}{text size}

\item{...}{other parameters}
}
\value{
A histogram and list of calculated values like confidence interval
}
\description{
This is a confidence interval placed on a histogram representing the function used.
}
