% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Itertions}

\item{x}{data input}

\item{fun}{function}

\item{alpha}{constant}

\item{cx}{Cx}

\item{...}{additional sttings for graph}
}
\description{
Bootstrap
}
